Moved PCOUNT variable to gene_score.py
