--- sccellfie/tests/__init__.py
@@ -1,1 +1,0 @@
-PCOUNT = 0. # Pseudocount used in function sccellfie.gene_score.gene_score()
